update for logut issue

+++ src/app/invictus-shared/top-nav/top-nav.component.ts
@@ -459,10 +459,12 @@ export class TopNavComponent implements OnInit, OnDestroy, AfterViewInit {
 						(result: any) => {
 							if (result.status === 'ok') {
 								this.deleteToken();
+								if (JSON.parse(localStorage.getItem('Prefix'))) {
 								var prefix1 = (JSON.parse(localStorage.getItem('Prefix')).pre);
 								localStorage.clear();
 								localStorage.setItem('Prefix', JSON.stringify({ pre: prefix1 }));
 								this.commonAPIService.setUserPrefix(prefix1);
+								}
 								const routeStore: RouteStore = new RouteStore();
 								routeStore.adm_no = '';
 								routeStore.login_id = '';
